requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions
openai
semantic_kernel
azure-monitor-opentelemetry-exporter==1.0.0b24
pydantic==2.8.2
